Moved popups from forms to components
