+++ app/javascript/components/AddPopup/index.js
@@ -1,0 +1,3 @@
+import AddPopup from "./AddPopup";
+
+export default AddPopup;
