+++ app/javascript/components/AddPopup/useStyles.js
@@ -1,0 +1,31 @@
+import { makeStyles } from "@material-ui/core/styles";
+
+const useStyles = makeStyles(() => ({
+  modal: {
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+    outline: 0,
+  },
+
+  root: {
+    width: 465,
+  },
+
+  actions: {
+    display: "flex",
+    justifyContent: "flex-end",
+  },
+
+  form: {
+    display: "flex",
+    flexDirection: "column",
+  },
+
+  loader: {
+    display: "flex",
+    justifyContent: "center",
+  },
+}));
+
+export default useStyles;
