+++ app/javascript/components/EditPopup/index.js
@@ -1,0 +1,3 @@
+import EditPopup from "./EditPopup";
+
+export default EditPopup;
